change encoding to PCM_FLOAT

+++ app/src/main/java/ru/luchemete/simplerecorder/audio/AudioDataReceivedListener.kt
@@ -1,5 +1,5 @@
 package ru.luchemete.simplerecorder.audio
 
 interface AudioDataReceivedListener {
-    fun onAudioDataReceived(data: ShortArray)
+    fun onAudioDataReceived(data: FloatArray)
 }
