+++ app/src/main/java/ru/luchemete/simplerecorder/audio/SingleRecord.kt
@@ -1,3 +1,3 @@
 package ru.luchemete.simplerecorder.audio
 
-data class SingleRecord(val data: ShortArray, val position: Int, val isRecording: Boolean)
+data class SingleRecord(val data: FloatArray, val position: Int, val isRecording: Boolean)
